Some cleanups after the last wave of changes
