+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 # be a sign or warning of (an imminent event, typically an unwelcome one).
 name = "presage"
-version = "0.3.0"
+version = "0.4.0"
 authors = ["Gabriel Féron <g@leirbag.net>"]
 edition = "2021"
 
@@ -59,15 +59,15 @@ tokio = { version = "1.0", features = [
 qr2term = { version = "0.2.2" }
 
 [features]
-default = ["sled-config-store"]
+default = ["sled-store"]
 quirks = []
-sled-config-store = ["sled"]
-secret-volatile-config-store = ["libsodium-sys", "secrets"]
-volatile-config-store = []
+sled-store = ["sled"]
+secret-volatile-store = ["libsodium-sys", "secrets"]
+volatile-store = []
 
-#[patch."https://github.com/whisperfish/libsignal-service-rs.git"]
-#libsignal-service = { path = "../libsignal-service-rs/libsignal-service" }
-#libsignal-service-hyper = { path = "../libsignal-service-rs/libsignal-service-hyper" }
+[patch."https://github.com/whisperfish/libsignal-service-rs.git"]
+libsignal-service = { path = "../libsignal-service-rs/libsignal-service" }
+libsignal-service-hyper = { path = "../libsignal-service-rs/libsignal-service-hyper" }
 
 [patch.crates-io]
-"curve25519-dalek" = { git = 'https://github.com/signalapp/curve25519-dalek', branch = 'lizard2' }
+curve25519-dalek = { git = "https://github.com/signalapp/curve25519-dalek", branch = "lizard2" }
