+++ src/lib.rs
@@ -1,21 +1,21 @@
 mod cache;
-mod config;
 mod errors;
 mod manager;
 mod proto;
+mod store;
 
-#[cfg(feature = "sled-config-store")]
-pub use config::sled::SledConfigStore;
+#[cfg(feature = "sled-store")]
+pub use store::sled::SledStore;
 
-#[cfg(feature = "secret-volatile-config-store")]
-pub use config::secret_volatile::SecretVolatileConfigStore;
-#[cfg(feature = "volatile-config-store")]
-pub use config::volatile::VolatileConfigStore;
+#[cfg(feature = "secret-volatile-store")]
+pub use store::secret_volatile::SecretVolatileStore;
+#[cfg(feature = "volatile-store")]
+pub use store::volatile::VolatileStore;
 
-pub use config::{ConfigStore, ContactsStore, MessageStore, StateStore, Thread};
 pub use errors::Error;
 pub use manager::{Confirmation, Linking, Manager, Registered, Registration, RegistrationOptions};
 pub use proto::ContentProto;
+pub use store::{ConfigStore, ContactsStore, MessageStore, StateStore, Thread};
 
 #[deprecated(note = "Please help use improve the prelude module instead")]
 pub use libsignal_service;
