+++ src/proto.rs
@@ -13,15 +13,17 @@ use libsignal_service::content::Metadata;
 use libsignal_service::prelude::Content;
 use libsignal_service::prelude::Uuid;
 
-impl AddressProto {
-    pub fn from_service(s: ServiceAddress) -> AddressProto {
+impl From<ServiceAddress> for AddressProto {
+    fn from(s: ServiceAddress) -> Self {
         AddressProto {
             uuid: s.uuid.map(|u| u.as_bytes().to_vec()),
             e164: s.e164(),
         }
     }
+}
 
-    pub fn into_service(self) -> ServiceAddress {
+impl Into<ServiceAddress> for AddressProto {
+    fn into(self) -> ServiceAddress {
         ServiceAddress {
             uuid: self
                 .uuid
@@ -32,10 +34,10 @@ impl AddressProto {
     }
 }
 
-impl MetadataProto {
-    pub fn from_metadata(m: Metadata) -> MetadataProto {
+impl From<Metadata> for MetadataProto {
+    fn from(m: Metadata) -> Self {
         MetadataProto {
-            address: Some(AddressProto::from_service(m.sender)),
+            address: Some(m.sender.into()),
             sender_device: m.sender_device.try_into().ok(),
             timestamp: m.timestamp.try_into().ok(),
             server_received_timestamp: None,
@@ -46,12 +48,14 @@ impl MetadataProto {
             destination_uuid: None,
         }
     }
+}
 
-    pub fn into_metadata(self) -> Metadata {
+impl Into<Metadata> for MetadataProto {
+    fn into(self) -> Metadata {
         Metadata {
             sender: self
                 .address
-                .map(|a| a.into_service())
+                .map(|a| a.into())
                 .unwrap_or_else(|| ServiceAddress {
                     uuid: None,
                     phonenumber: None,
@@ -78,53 +82,23 @@ pub struct ContentProto {
     content: crate::prelude::proto::Content,
 }
 
-impl ContentProto {
-    pub fn from_content(c: Content) -> ContentProto {
+impl From<Content> for ContentProto {
+    fn from(c: Content) -> Self {
         ContentProto {
-            metadata: MetadataProto::from_metadata(c.metadata),
+            metadata: c.metadata.into(),
             content: c.body.into_proto(),
         }
     }
-
-    pub fn into_content(self) -> Content {
-        content_from_proto(self.content, self.metadata.into_metadata())
-            .expect("Content to have at least one type")
-    }
 }
 
-/// TODO: From  libsignal_service
-/// Converts a proto::Content into a public Content, including metadata.
-pub(crate) fn content_from_proto(
-    p: libsignal_service::proto::Content,
-    metadata: Metadata,
-) -> Option<libsignal_service::content::Content> {
-    // The Java version also assumes only one content type at a time.
-    // It's a bit sad that we cannot really match here, we've got no
-    // r#type() method.
-    // Allow the manual map (if let Some -> option.map(||)), because it
-    // reduces the git diff when more types would be added.
-    #[allow(clippy::manual_map)]
-    if let Some(msg) = p.data_message {
-        Some(libsignal_service::content::Content::from_body(
-            msg, metadata,
-        ))
-    } else if let Some(msg) = p.sync_message {
-        Some(libsignal_service::content::Content::from_body(
-            msg, metadata,
-        ))
-    } else if let Some(msg) = p.call_message {
-        Some(libsignal_service::content::Content::from_body(
-            msg, metadata,
-        ))
-    } else if let Some(msg) = p.receipt_message {
-        Some(libsignal_service::content::Content::from_body(
-            msg, metadata,
-        ))
-    } else if let Some(msg) = p.typing_message {
-        Some(libsignal_service::content::Content::from_body(
-            msg, metadata,
-        ))
-    } else {
-        None
+#[derive(thiserror::Error, Debug)]
+enum ContentProtoError {}
+
+impl TryInto<Content> for ContentProto {
+    type Error = crate::Error;
+
+    fn try_into(self) -> Result<Content, Self::Error> {
+        Content::from_proto(self.content, self.metadata.into())
+            .ok_or(crate::Error::ContentMissingMessage)
     }
 }
