+++ src/store/volatile.rs
@@ -1,6 +1,6 @@
 use std::{
     collections::HashMap,
-    sync::{Arc, Mutex, RwLock},
+    sync::{Arc, RwLock},
 };
 
 use async_trait::async_trait;
@@ -8,52 +8,36 @@ use libsignal_service::{
     models::Contact,
     prelude::{
         protocol::{
-            Context, Direction, IdentityKey, IdentityKeyPair, IdentityKeyStore, PreKeyRecord,
-            PreKeyStore, ProtocolAddress, SessionRecord, SessionStore, SessionStoreExt,
-            SignalProtocolError, SignedPreKeyRecord, SignedPreKeyStore,
+            Context, Direction, IdentityKey, IdentityKeyPair, IdentityKeyStore, PreKeyId,
+            PreKeyRecord, PreKeyStore, ProtocolAddress, SessionRecord, SessionStore,
+            SessionStoreExt, SignalProtocolError, SignedPreKeyId, SignedPreKeyRecord,
+            SignedPreKeyStore,
         },
-        Content, Uuid,
+        Uuid,
     },
     push_service::DEFAULT_DEVICE_ID,
-    ServiceAddress,
 };
 use log::{trace, warn};
-use secrets::{SecretBox, SecretVec};
 
-use super::{ConfigStore, ContactsStore, StateStore, Thread};
-use crate::{manager::Registered, Error, MessageStore};
-
-// - SecretId adds the Default trait to SecretBox<u32>
-#[derive(Debug, Clone)]
-struct SecretId(SecretBox<u32>);
-
-impl Default for SecretId {
-    fn default() -> Self {
-        SecretId(SecretBox::new(|s| *s = 0u32))
-    }
-}
+use super::{ConfigStore, ContactsStore, StateStore};
+use crate::{manager::Registered, Error};
 
 #[derive(Default, Debug, Clone)]
-pub struct SecretVolatileConfigStore {
-    // `Cell<u8>` cannot be shared between threads safely (part of SecretBox/SecretVec)
-    // - therefore wrapped within a Mutex
-    // - to be able to derive the Clone trait Arc is needed
-    pre_keys_offset_id: Arc<Mutex<SecretId>>,
-    next_signed_pre_key_id: Arc<Mutex<SecretId>>,
+pub struct VolatileStore {
+    pre_keys_offset_id: u32,
+    next_signed_pre_key_id: u32,
 
-    pre_keys: Arc<RwLock<HashMap<u32, Mutex<SecretVec<u8>>>>>,
-    signed_pre_keys: Arc<RwLock<HashMap<u32, Mutex<SecretVec<u8>>>>>,
+    pre_keys: HashMap<PreKeyId, Vec<u8>>,
+    signed_pre_keys: HashMap<SignedPreKeyId, Vec<u8>>,
 
     // XXX: we need interior mutability + Sync until we fix the trait definition to use &mut self in libsignal-service
-    sessions: Arc<RwLock<HashMap<String, Mutex<SecretVec<u8>>>>>,
+    sessions: Arc<RwLock<HashMap<String, Vec<u8>>>>,
+    identities: HashMap<ProtocolAddress, Vec<u8>>,
 
-    identities: Arc<RwLock<HashMap<ProtocolAddress, Mutex<SecretVec<u8>>>>>,
-    registration: Arc<Mutex<Option<SecretVec<u8>>>>,
+    registration: Option<Vec<u8>>,
 }
-// todo: to erase or zero out secrets passed into the secret store upstream refactoring is necessary,
-// parameters for example `id` for `set_pre_keys_offset_id` must be passed as &mut for the secrets crate to zero out the value.
 
-impl SecretVolatileConfigStore {
+impl VolatileStore {
     fn session_key(&self, addr: &ProtocolAddress) -> String {
         format!("session-{}", addr)
     }
@@ -63,54 +47,43 @@ impl SecretVolatileConfigStore {
     }
 }
 
-impl StateStore<Registered> for SecretVolatileConfigStore {
+impl StateStore<Registered> for VolatileStore {
     fn load_state(&self) -> Result<Registered, Error> {
-        let d = self.registration.try_lock().expect("poisoned mutex");
-        let data = d.as_ref().ok_or(Error::NotYetRegisteredError)?;
-        let x = serde_json::from_slice(&*data.borrow()).map_err(Error::from);
-        x
+        let data = self
+            .registration
+            .as_ref()
+            .ok_or(Error::NotYetRegisteredError)?;
+        serde_json::from_slice(data).map_err(Error::from)
     }
 
     fn save_state(&mut self, state: &Registered) -> Result<(), Error> {
-        let mut data = serde_json::to_vec(state)?;
-        self.registration = Arc::new(Mutex::new(Some(SecretVec::from(&mut data[..]))));
+        let data = serde_json::to_vec(state)?;
+        self.registration = Some(data);
         Ok(())
     }
 }
 
-impl ConfigStore for SecretVolatileConfigStore {
+impl ConfigStore for VolatileStore {
     fn pre_keys_offset_id(&self) -> Result<u32, Error> {
-        let d = self.pre_keys_offset_id.try_lock().expect("poisoned mutex");
-        let x = *d.0.borrow();
-        Ok(x)
+        Ok(self.pre_keys_offset_id)
     }
 
     fn set_pre_keys_offset_id(&mut self, id: u32) -> Result<(), Error> {
-        let mut d = self.pre_keys_offset_id.try_lock().expect("poisoned mutex");
-        *d.0.borrow_mut() = id;
+        self.pre_keys_offset_id = id;
         Ok(())
     }
 
     fn next_signed_pre_key_id(&self) -> Result<u32, Error> {
-        let d = self
-            .next_signed_pre_key_id
-            .try_lock()
-            .expect("poisoned mutex");
-        let x = *d.0.borrow();
-        Ok(x)
+        Ok(self.next_signed_pre_key_id)
     }
 
     fn set_next_signed_pre_key_id(&mut self, id: u32) -> Result<(), Error> {
-        let mut d = self
-            .next_signed_pre_key_id
-            .try_lock()
-            .expect("poisoned mutex");
-        *d.0.borrow_mut() = id;
+        self.next_signed_pre_key_id = id;
         Ok(())
     }
 }
 
-impl ContactsStore for SecretVolatileConfigStore {
+impl ContactsStore for VolatileStore {
     fn save_contacts(&mut self, _: impl Iterator<Item = Contact>) -> Result<(), Error> {
         warn!("contacts are not saved when using volatile storage.");
         Ok(())
@@ -127,107 +100,68 @@ impl ContactsStore for SecretVolatileConfigStore {
     }
 }
 
-impl MessageStore for SecretVolatileConfigStore {
-    type MessagesIter = std::vec::IntoIter<Content>;
-    fn save_message(
-        &mut self,
-        _message: libsignal_service::prelude::Content,
-        _: Option<impl Into<ServiceAddress>>,
-    ) -> Result<(), Error> {
-        warn!("messages are not saved when using volatile storage.");
-        Ok(())
-    }
-
-    fn message(
-        &self,
-        _thread: &Thread,
-        _timestamp: u64,
-    ) -> Result<Option<libsignal_service::prelude::Content>, Error> {
-        warn!("messages are not saved when using volatile storage.");
-        Ok(None)
-    }
-
-    fn messages(&self, _thread: &Thread, _from: Option<u64>) -> Result<Self::MessagesIter, Error> {
-        warn!("messages are not saved when using volatile storage.");
-        Ok(vec![].into_iter())
-    }
-}
-
 #[async_trait(?Send)]
-impl PreKeyStore for SecretVolatileConfigStore {
+impl PreKeyStore for VolatileStore {
     async fn get_pre_key(
         &self,
-        prekey_id: u32,
+        prekey_id: PreKeyId,
         _ctx: Context,
     ) -> Result<PreKeyRecord, SignalProtocolError> {
-        let buf = self.pre_keys.try_read().expect("poisoned mutex");
-        let b = buf
+        let buf = self
+            .pre_keys
             .get(&prekey_id)
             .ok_or(SignalProtocolError::InvalidPreKeyId)?;
-        let x = PreKeyRecord::deserialize(&(*b.try_lock().unwrap()).borrow());
-        x
+        PreKeyRecord::deserialize(buf)
     }
 
     async fn save_pre_key(
         &mut self,
-        prekey_id: u32,
+        prekey_id: PreKeyId,
         record: &PreKeyRecord,
         _ctx: Context,
     ) -> Result<(), SignalProtocolError> {
-        self.pre_keys.try_write().expect("poisoned mutex").insert(
-            prekey_id,
-            Mutex::new(SecretVec::from(&mut record.serialize()?[..])),
-        );
+        self.pre_keys.insert(prekey_id, record.serialize()?);
         Ok(())
     }
 
     async fn remove_pre_key(
         &mut self,
-        prekey_id: u32,
+        prekey_id: PreKeyId,
         _ctx: Context,
     ) -> Result<(), SignalProtocolError> {
-        self.pre_keys
-            .try_write()
-            .expect("poisoned mutex")
-            .remove(&prekey_id);
+        self.pre_keys.remove(&prekey_id);
         Ok(())
     }
 }
 
 #[async_trait(?Send)]
-impl SignedPreKeyStore for SecretVolatileConfigStore {
+impl SignedPreKeyStore for VolatileStore {
     async fn get_signed_pre_key(
         &self,
-        signed_prekey_id: u32,
+        signed_prekey_id: SignedPreKeyId,
         _ctx: Context,
     ) -> Result<SignedPreKeyRecord, SignalProtocolError> {
-        let buf = self.signed_pre_keys.try_read().expect("poisoned mutex");
-        let b = buf
+        let buf = self
+            .signed_pre_keys
             .get(&signed_prekey_id)
             .ok_or(SignalProtocolError::InvalidSignedPreKeyId)?;
-        let x = SignedPreKeyRecord::deserialize(&(*b.try_lock().unwrap()).borrow());
-        x
+        SignedPreKeyRecord::deserialize(buf)
     }
 
     async fn save_signed_pre_key(
         &mut self,
-        signed_prekey_id: u32,
+        signed_prekey_id: SignedPreKeyId,
         record: &SignedPreKeyRecord,
         _ctx: Context,
     ) -> Result<(), SignalProtocolError> {
         self.signed_pre_keys
-            .try_write()
-            .expect("poisoned mutex")
-            .insert(
-                signed_prekey_id,
-                Mutex::new(SecretVec::from(&mut record.serialize()?[..])),
-            );
+            .insert(signed_prekey_id, record.serialize()?);
         Ok(())
     }
 }
 
 #[async_trait(?Send)]
-impl SessionStore for SecretVolatileConfigStore {
+impl SessionStore for VolatileStore {
     async fn load_session(
         &self,
         address: &ProtocolAddress,
@@ -237,8 +171,7 @@ impl SessionStore for SecretVolatileConfigStore {
         let key = self.session_key(address);
         let buf = db.get(&key);
 
-        buf.map(|buf| SessionRecord::deserialize(&(*buf.try_lock().unwrap()).borrow()))
-            .transpose()
+        buf.map(|buf| SessionRecord::deserialize(buf)).transpose()
     }
 
     async fn store_session(
@@ -248,17 +181,16 @@ impl SessionStore for SecretVolatileConfigStore {
         _ctx: Context,
     ) -> Result<(), SignalProtocolError> {
         let key = self.session_key(address);
-        let mut data = record.serialize()?;
         self.sessions
             .try_write()
             .expect("poisoned mutex")
-            .insert(key, Mutex::new(SecretVec::from(&mut data[..])));
+            .insert(key, record.serialize()?);
         Ok(())
     }
 }
 
 #[async_trait]
-impl SessionStoreExt for SecretVolatileConfigStore {
+impl SessionStoreExt for VolatileStore {
     async fn get_sub_device_sessions(&self, name: &str) -> Result<Vec<u32>, SignalProtocolError> {
         let session_prefix = self.session_prefix(name);
         log::info!("get_sub_device_sessions: session_prefix={}", session_prefix);
@@ -295,7 +227,7 @@ impl SessionStoreExt for SecretVolatileConfigStore {
 }
 
 #[async_trait(?Send)]
-impl IdentityKeyStore for SecretVolatileConfigStore {
+impl IdentityKeyStore for VolatileStore {
     async fn get_identity_key_pair(
         &self,
         _ctx: Context,
@@ -323,10 +255,8 @@ impl IdentityKeyStore for SecretVolatileConfigStore {
         identity_key: &IdentityKey,
         _ctx: Context,
     ) -> Result<bool, SignalProtocolError> {
-        self.identities.try_write().expect("poisoned mutex").insert(
-            address.clone(),
-            Mutex::new(SecretVec::from(&mut identity_key.serialize().to_vec()[..])),
-        );
+        self.identities
+            .insert(address.clone(), identity_key.serialize().to_vec());
         Ok(false)
     }
 
@@ -337,20 +267,13 @@ impl IdentityKeyStore for SecretVolatileConfigStore {
         _direction: Direction,
         _ctx: Context,
     ) -> Result<bool, SignalProtocolError> {
-        match self
-            .identities
-            .try_read()
-            .expect("poisoned mutex")
-            .get(address)
-        {
+        match self.identities.get(address) {
             None => {
                 // when we encounter a new identity, we trust it by default
                 warn!("trusting new identity {:?}", address);
                 Ok(true)
             }
-            Some(contents) => Ok(
-                &IdentityKey::decode(&(*contents.try_lock().unwrap()).borrow())? == identity_key,
-            ),
+            Some(contents) => Ok(&IdentityKey::decode(contents)? == identity_key),
         }
     }
 
@@ -359,8 +282,7 @@ impl IdentityKeyStore for SecretVolatileConfigStore {
         address: &ProtocolAddress,
         _ctx: Context,
     ) -> Result<Option<IdentityKey>, SignalProtocolError> {
-        let identities = self.identities.try_write().expect("poisoned mutex");
-        let buf = identities.get(address);
-        Ok(buf.map(|b| IdentityKey::decode(&(*b.try_lock().unwrap()).borrow()).unwrap()))
+        let buf = self.identities.get(address);
+        Ok(buf.map(|b| IdentityKey::decode(b).unwrap()))
     }
 }
